refactor(simulation): change accumulative probability to range

+++ src/main/kotlin/markov/simulation/Moving.kt
@@ -4,7 +4,7 @@ import markov.map.Position
 
 @JvmInline
 value class Moving(val probabilities: Map<Position, Action>) {
-    fun nextPosition(currentPosition: Position, probability: Double): Position {
+    fun nextPosition(currentPosition: Position, probability: Int): Position {
         val actions =
             probabilities[currentPosition] ?: throw IllegalStateException("$currentPosition 에 대한 정보를 찾을 수 없습니다")
         return currentPosition.next(actions.chooseAction(probability))
